chore:get single product

+++ controllers/deleteProduct.js
@@ -5,12 +5,12 @@ const deleteProduct = async (req, res, next) => {
   try {
     // Retrieve the product from the database
     const product = await ProductSchema.findOne({ _id: req.params.productId });
-
+if (!product) return res.status(404).send({msg:"item with the given id not found!"})
     // Retrieve the product_id from the image array
     const productIds = product.image.map(image => image.product_id);
 
     // Delete the images from Cloudinary
-    // await cloudinary.api.delete_resources(productIds);
+    await cloudinary.api.delete_resources(productIds);
       await cloudinary.api.delete_resources(productIds)
 
     // Delete the product from the database
